% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_station.R
\name{ace_station}
\alias{ace_station}
\title{Create Objects Representing Water Monitoring Stations}
\usage{
ace_station(name, code, data_columns, depths = c(), data_skiprows,
  url = NULL)
}
\arguments{
\item{name}{The name of the station (e.g. "Fremont Bridge)}

\item{code}{Water monitoring station code (e.g. "FBLW")}

\item{data_columns}{Vector of names of columns in the data}

\item{depths}{Vector of depths, in feet, for the station's measurements}

\item{data_skiprows}{Number of rows to skip at the beginning of data file}

\item{url}{Station data url. By default, one will be created using the
standard url scheme and the monitoring station code.}
}
\value{
An \code{ace_station} object
}
\description{
\code{ace_station} constructs \code{ace_station} objects, which store
information about water monitoring stations
}
\examples{
my_station <- ace_station(
    name = "Fremont Bridge",
    code = "FBLW",
    data_columns = c("Time", "SaltA", "SaltB", "SaltC", "TempA", "TempB", "TempC"),
    depths = c("A" = 18, "B" = 31, "C" = 40),
    data_skiprows = 5
)

}
