% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acewater.R
\docType{package}
\name{acewater-package}
\alias{acewater}
\alias{acewater-package}
\title{acewater: Download Water Conditions for the Lake Washington Ship Canal}
\description{
This package returns tidy data from the US Army Corps of Engineers'
water quality data for the Lake Washington Ship Canal.
}
\note{
See \url{http://www.nwd-wc.usace.army.mil/nws/hh/www/index.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/acewater}
  \item Report bugs at \url{https://github.com/briandconnelly/acewater/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net}

}
