% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ace_conditions.R
\name{autoplot.ace_conditions}
\alias{autoplot.ace_conditions}
\title{Plot Water Conditions using ggplot2}
\usage{
autoplot.ace_conditions(object, ...)
}
\arguments{
\item{object}{An \code{ace_conditions} object}

\item{...}{Additional arguments (not used)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
\code{autoplot.ace_conditions} creates a basic plot of water conditions using
the ggplot2 package (if installed). Additional formatting can be done by
adding to the object that is returned.
}
\examples{
\dontrun{
library(acewater)
library(ggplot2)

cUBLW <- get_water_conditions("UBLW")
autoplot(cUBLW)
}
}
