% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ace_station.R
\name{autoplot.ace_station}
\alias{autoplot.ace_station}
\title{Plot a Station's Water Conditions using ggplot2}
\usage{
autoplot.ace_station(object, ...)
}
\arguments{
\item{object}{An \code{ace_station} object}

\item{...}{Additional arguments (not used)}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
\code{autoplot.ace_station} retrieves the latest water conditions and
creates a basic plot using the ggplot2 package (if installed).
}
\examples{
\dontrun{
library(acewater)
library(ggplot2)

s1 <- get_station(station = "UBLW", project = "LKW")
autoplot(s1)
}
}
