% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_project.R
\name{find_project}
\alias{find_project}
\title{Find the Project/Basin Associated with a Monitoring Station}
\usage{
find_project(station)
}
\arguments{
\item{station}{Water monitoring station code (e.g. "FBLW")}
}
\value{
A character array containing the project code (e.g. "LKW"). If the
station is unknown, \code{NULL} is returned.
}
\description{
Find the Project/Basin Associated with a Monitoring Station
}
\examples{
library(acewater)
find_project("FBLW")

}
