% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station.R
\name{get_station}
\alias{get_station}
\title{Get Information about a Monitoring Station}
\usage{
get_station(project, code, ...)
}
\arguments{
\item{project}{Project/basin code (e.g. "LKW")}

\item{code}{Water monitoring station code (e.g. "FBLW")}

\item{...}{Additional arguments (not supported)}
}
\value{
A list containing information about a monitoring station
}
\description{
Get Information about a Monitoring Station
}
\examples{
library(acewater)
s <- get_station(project = "LKW", code = "FBLW")

}
