% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_name.R
\name{station_name}
\alias{station_name}
\title{Retrieve Information about a Monitoring Station}
\usage{
station_name(code)
}
\arguments{
\item{code}{Water monitoring station code (e.g. "FBLW")}
}
\description{
Retrieve Information about a Monitoring Station
}
\examples{
library(acewater)

station_name("FBLW")

}
