% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_station.R
\name{ace_station}
\alias{ace_station}
\title{Create Object Representing a Water Monitoring Station}
\usage{
ace_station(name, code, project, data_columns, data_units = c(),
  depths = c(), data_skiprows, data_url = NULL, latitude = NULL,
  longitude = NULL, web_url = NULL)
}
\arguments{
\item{name}{The name of the station (e.g. "Fremont Bridge")}

\item{code}{Water monitoring station code (e.g. "FBLW")}

\item{project}{Code for project (e.g. "LKW" for Lake Washington Ship Canal)}

\item{data_columns}{Vector of names of columns in the data}

\item{data_units}{Vector of units corresponding to each entry in data_columns}

\item{depths}{Vector of depths, in feet, for the station's measurements}

\item{data_skiprows}{Number of rows to skip at the beginning of data file}

\item{data_url}{Station data url. By default, one will be created using the
standard url scheme and the monitoring station code.}

\item{latitude}{Station's latitude}

\item{longitude}{Station's longitude}

\item{web_url}{URL providing information about the station}
}
\value{
An \code{ace_station} object
}
\description{
\code{ace_station} constructs \code{ace_station} objects, which store
information about water monitoring stations
}
\examples{
my_station <- ace_station(
    name = "Fremont Bridge",
    code = "FBLW",
    data_columns = c("Time", "SaltA", "SaltB", "SaltC", "TempA", "TempB", "TempC"),
    depths = c("A" = 18, "B" = 31, "C" = 40),
    data_skiprows = 5
)

}
