% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_conditions.R
\name{get_water_conditions}
\alias{get_water_conditions}
\title{Get Water Conditions for a Monitoring Station}
\usage{
get_water_conditions(station, ...)
}
\arguments{
\item{station}{A list containing station information. See \code{\link{get_station}}.}

\item{...}{Additional arguments (not supported)}
}
\value{
A tibble
}
\description{
Get Water Conditions for a Monitoring Station
}
\examples{
library(acewater)
s <- get_station(code = "FBLW")
get_water_conditions(station = s)

}
