% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_conditions.R
\name{get_water_conditions}
\alias{get_water_conditions}
\title{Get Water Conditions for a Monitoring Station}
\usage{
get_water_conditions(station, na.rm = FALSE, ...)
}
\arguments{
\item{station}{Water monitoring station code (e.g. "FBLW").}

\item{na.rm}{A logical value indicating whether \code{NA} values should be removed from the results (default: \code{FALSE}).}

\item{...}{Additional arguments (not supported)}
}
\value{
A tibble
}
\description{
Get Water Conditions for a Monitoring Station
}
